backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
requests
langchain
langchain-openai
langgraph
langchain-community
tavily-python
google-genai
apscheduler
markdown
